python-backend/requirements.txt
langchain
langchain-community
langchain_huggingface
langchain-text-splitters
python-dotenv
bs4
sentence_transformers
langchain-cli
langchain_chroma
langchain-groq
langchain-astradb
langchain-unstructured
langserve
streamlit
fastapi
pydantic
uvicorn
pandas
duckdb
duckduckgo_search==5.3.1b1
SQLAlchemy
validators==0.28.1
unstructured
pytube
numexpr
huggingface_hub
uuid
googlemaps